+++ CHANGES.md
@@ -37,6 +37,10 @@
 - [ADD] PyPI に `hiusi` を登録する GitHub Actions `pypi-publish.yml` を追加する
   - バージョンが `-canary.X` は `.devX` 形式に変換される
   - @voluntas
+- [ADD] 実験的コマンドとして rtmp-inbound-endpoint を追加する
+  - RTMP サーバーとして受信したストリームを出力ファイル（.mp4）に保存するためのコマンド
+  - `--experimental(-x)` フラグが指定されている場合にのみ利用可能となる
+  - @sile
 - [ADD] 実験的コマンドとして rtmp-outbound-endpoint を追加する
   - 入力ファイル（.mp4 or .webm）を RTMP サーバーとして配信するためのコマンド
   - `--experimental(-x)` フラグが指定されている場合にのみ利用可能となる